requirements.txt
Django>=5.0
django-unfold
pillow  # Necessário para uploads de imagens (ex: logo no orçamento)